feat: Cambio Tipo dato Integer x Long, models TC5,TC10,TC48,swdmplog

+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/incomingvisa/IncomingVisaTC10TCR0Det.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = false)
 public class IncomingVisaTC10TCR0Det {
-	Integer secuenciaIncoming;
+	Long secuenciaIncoming;
 	Date fechaProceso;
 	String transactionCode;
 	String descTransactionCode;
